refactor: rename cli_tools package to cli
