+++ dfode_kit/cli_tools/__init__.py
@@ -1,0 +1,3 @@
+"""Compatibility shims for the renamed ``dfode_kit.cli_tools`` package."""
+
+from dfode_kit.cli import *  # noqa: F401,F403
