+++ dfode_kit/cli_tools/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility shims for :mod:`dfode_kit.cli.commands`."""
