+++ dfode_kit/cli_tools/commands/sample.py
@@ -1,39 +1,3 @@
-import argparse
+"""Compatibility shim for :mod:`dfode_kit.cli.commands.sample`."""
 
-
-def add_command_parser(subparsers):
-    sample_parser = subparsers.add_parser('sample', help='Perform sampling.')
-
-    sample_parser.add_argument(
-        '--mech',
-        required=True,
-        type=str,
-        help='Path to the mechanism file.'
-    )
-    sample_parser.add_argument(
-        '--case',
-        required=True,
-        type=str,
-        help='Root directory containing data.'
-    )
-    sample_parser.add_argument(
-        '--save',
-        required=True,
-        type=str,
-        help='Path where the HDF5 file will be saved.'
-    )
-    sample_parser.add_argument(
-        '--include_mesh',
-        action='store_true',
-        help='Include mesh data in the HDF5 file.'
-    )
-
-
-def handle_command(args):
-    from dfode_kit.data_operations.h5_kit import touch_h5
-    from dfode_kit.df_interface.sample_case import df_to_h5
-
-    print('Handling sample command')
-    df_to_h5(args.case, args.mech, args.save, include_mesh=args.include_mesh)
-    print()
-    touch_h5(args.save)
+from dfode_kit.cli.commands.sample import *  # noqa: F401,F403
