+++ dfode_kit/cli_tools/commands/train.py
@@ -1,28 +1,3 @@
-def add_command_parser(subparsers):
-    train_parser = subparsers.add_parser('train', help='Train the model.')
-    train_parser.add_argument(
-        '--mech',
-        required=True,
-        type=str,
-        help='Path to the YAML mechanism file.'
-    )
-    train_parser.add_argument(
-        '--source_file',
-        required=True,
-        type=str,
-        help='Path to the source NUMPY file. (With source data and labeled data)'
-    )
-    train_parser.add_argument(
-        '--output_path',
-        required=True,
-        type=str,
-        help='Path to the output model.'
-    )
+"""Compatibility shim for :mod:`dfode_kit.cli.commands.train`."""
 
-
-def handle_command(args):
-    from dfode_kit.dfode_core.train.train import train
-
-    print('Handling train command')
-    train(args.mech, args.source_file, args.output_path)
-    print(f'Saved Model to {args.output_path}')
+from dfode_kit.cli.commands.train import *  # noqa: F401,F403
