+++ docs/architecture.md
@@ -2,7 +2,7 @@
 
 ## Current repository structure
 
-- `dfode_kit/cli_tools/`: CLI entrypoints and subcommands
+- `dfode_kit/cli/`: CLI entrypoints and subcommands
 - `dfode_kit/data_operations/`: dataset I/O, contracts, labeling, augmentation, integration utilities
 - `dfode_kit/dfode_core/`: models, training, registries, preprocessing
 - `dfode_kit/df_interface/`: DeepFlame/OpenFOAM-facing helpers
