+++ pyproject.toml
@@ -19,7 +19,7 @@ dev = [
 ]
 
 [project.scripts]
-dfode-kit = "dfode_kit.cli_tools.main:main"
+dfode-kit = "dfode_kit.cli.main:main"
 
 [tool.setuptools.packages.find]
-include = ["dfode_kit"]
+include = ["dfode_kit", "dfode_kit.*"]
